requirements.txt
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-MySQLdb==0.2.0
Flask-pymysql==0.2.3
Flask-Script==2.0.6
flask-sqlacodegen==1.1.6.1
Flask-SQLAlchemy==2.4.0
idna==2.8
inflect==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
requests==2.21.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2
uwsgi
